abstracting parser because it isn't necessarily console bound
