+++ Conways/Program.cs
@@ -6,7 +6,7 @@
     {
       var userInput = new ConsoleInput();
       var dimensions = userInput.GetDimensions();
-      var indexes = userInput.GetStartingState(dimensions);
+      var indexes = userInput.GetValidIndexes(dimensions);
       var world = new World(dimensions.Item1, dimensions.Item2, indexes);
       var renderer = new ConsoleRenderer();
       Simulation.Run(renderer, world);
